mhr-api/pyproject.toml
[tool.poetry]
name = "mhr-api"
version = "2.0.0"
description = ""
authors = ["dlovett <doug@daxiom.com>"]
license = "BSD 3"
readme = "README.md"
packages = [{include = "mhr_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
Flask-Cors = "^4.0.1"
flask-migrate = "^4.0.7"
Flask-Pydantic = "^0.12.0"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
launchdarkly-server-sdk = "^8.1.1"
pycountry = "^24.6.1"
email-validator= "1.3.1"
phonenumbers = "8.13.39"
notifications-python-client = "7.0.0"
beautifulsoup4 = "4.12.3"
sqlalchemy-continuum = "^1.4.2"
pg8000 = "^1.31.2"
cachecontrol = "^0.14.0"
structured-logging = { git = "https://github.com/bcgov/sbc-connect-common.git", subdirectory = "python/structured-logging", branch = "main" }
oauthlib = "^3.2.2"
Jinja2 = "^3.1.2"
google-api-core = "^2.11.0"
google-auth-oauthlib = "^1.0.0"
google-auth = "^2.16.2"
google-cloud-core = "^2.3.2"
google-cloud-pubsub = "^2.15.0"
google-cloud-storage = "^2.7.0"
googleapis-common-protos = "^1.58.0"
grpc-google-iam-v1 = "^0.12.6"
grpcio-status = "^1.51.3"
PyPDF2 = "^3.0.1"
pytz = "^2022.7.1"
alembic-utils = "^0.8.3"
datedelta = "^1.4"
flask-jwt-oidc = "^0.7.0"
registry-schemas =  { git = "https://github.com/bcgov/registry-schemas.git", tag = "1.8.11" }
flask-babel = "^4.0.0"


[tool.poetry.group.test.dependencies]
psycopg2-binary = "^2.9.9"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
requests-mock = "^1.12.1"
faker = "^25.9.2"
freezegun = "^1.5.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pylint = "^3.2.5"
bandit = "^1.7.9"
flake8-pyproject = "^1.2.3"
isort = "^5.13.2"

[tool.bandit]
exclude_dirs = [".venv","tests"]
skips = ["B104"]

[tool.flake8]
ignore = ["F401","E402", "Q000", "E203", "W503"]
exclude = [
    ".venv",
    ".git",
    ".history",
    "devops",
    "tests",
    "database",
    "*migrations*",
]
per-file-ignores = [
    "__init__.py:F401",
    "*.py:B902"
]
max-line-length = 120
docstring-min-length=10
count = true

[tool.black]
target-version =  ["py310", "py311", "py312"]
line-length = 120
include = '\.pyi?$'
force-exclude = '''
/(
  # The following are specific to Black, you probably don't want those.
  migrations
  | devops
  | tests
  | database
  | .history
)/
'''

[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true
skip_glob = ["migrations", "devops", "tests", "src/database/*"]

[tool.pylint.main]
fail-under = 10
max-line-length = 120
ignore = [ "migrations", "devops", "tests", "database"]
ignore-patterns = ["^\\.#"]
ignored-modules= ["flask_sqlalchemy", "sqlalchemy", "SQLAlchemy" , "alembic", "scoped_session"]
ignored-classes= "scoped_session"
ignore-long-lines = "^\\s*(# )?<?https?://\\S+>?$"
extension-pkg-whitelist = "pydantic"
notes = ["FIXME","XXX","TODO"]
overgeneral-exceptions = ["builtins.BaseException", "builtins.Exception"]
confidence = ["HIGH", "CONTROL_FLOW", "INFERENCE", "INFERENCE_FAILURE", "UNDEFINED"]
disable = "C0209,C0301,W0511,W0613,W0703,W1514,W1203,R0801,R0902,R0903,R0911,R0401,R1705,R1718,W3101"
argument-naming-style = "snake_case"
attr-naming-style = "snake_case"
class-attribute-naming-style = "any"
class-const-naming-style = "UPPER_CASE"
class-naming-style = "PascalCase"
const-naming-style = "UPPER_CASE"
function-naming-style = "snake_case"
inlinevar-naming-style = "any"
method-naming-style = "snake_case"
module-naming-style = "any"
variable-naming-style = "snake_case"
docstring-min-length = -1
good-names = ["i", "j", "k", "ex", "Run", "_"]
bad-names = ["foo", "bar", "baz", "toto", "tutu", "tata"]
defining-attr-methods = ["__init__", "__new__", "setUp", "asyncSetUp", "__post_init__"]
exclude-protected = ["_asdict", "_fields", "_replace", "_source", "_make", "os._exit"]
valid-classmethod-first-arg = ["cls"]
valid-metaclass-classmethod-first-arg = ["mcs"]

[tool.pytest.ini_options]
minversion = "2.0"
testpaths = [
   "tests",
]
# addopts = "--verbose --strict -p no:warnings --cov=src --cov-report html:htmlcov --cov-report xml:coverage.xml"
addopts = "--verbose --strict -p no:warnings"
python_files = [
   "test*.py"
]
norecursedirs = [
   ".git", ".tox", "venv*", "requirements*", "build", "database*",
]
log_cli = true
log_cli_level = "1"
filterwarnings = [
   "ignore::UserWarning"
]
markers = [
   "slow",
   "serial",
]

[tool.coverage.run]
branch = true
source = [
   "src/mhr_api",
]
omit = [
    "wsgi.py",
    "gunicorn_config.py"
]

[tool.coverage.report]
exclude_lines = [
   "pragma: no cover",
   "from",
   "import",
   "def __repr__",
   "if self.debug:",
   "if settings.DEBUG",
   "raise AssertionError",
   "raise NotImplementedError",
   "if 0:",
   'if __name__ == "__main__":',
]

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"